+++ src/main/java/com/cryptomanager/controllers/TransactionsController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/transactions-history")
+@RequestMapping("5/transactions-history")
 public class TransactionsController {
     private final TransactionsService transactionsService;
     private final LoginRepository loginRepository;
@@ -24,23 +24,6 @@ public class TransactionsController {
         this.loginRepository = loginRepository;
     }
 
-    @GetMapping("/get-full-history-ADMIN")
-    public ResponseEntity<String> getTransactionsHistory(@Parameter(description = "Transaction type", schema = @Schema(allowableValues = {"BUY", "SELL", "CONVERSION", "ALL"})) @RequestParam String transactionType) {
-        try {
-            return ResponseEntity.ok(transactionsService.getTransactionHistory(transactionType));
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
-        }
-    }
-
-    @GetMapping("/get-history-by-ID-ADMIN")
-    public ResponseEntity<String> getTransactionsHistoryByID(@RequestParam String userID, @Parameter(description = "Transaction type", schema = @Schema(allowableValues = {"BUY", "SELL", "CONVERSION", "ALL"})) @RequestParam String transactionType) {
-        try {
-            return ResponseEntity.ok(transactionsService.getTransactionHistoryByID(transactionType, userID));
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
-        }
-    }
     @GetMapping("/get-own-history")
     public ResponseEntity<String> getLoggedClientHistory(@Parameter(description = "Transaction type", schema = @Schema(allowableValues = {"BUY", "SELL", "CONVERSION", "ALL"})) @RequestParam String transactionType) {
         try {
